roland-modela: Test update motor mount
Broken design, cables+ESC does not have a way to come through.
Alternatives for fixing:
* have a cable channel that goes down to underside and out
* twist the motor 90 deg, open up one of the sides again all the way down
* open up a hole big enough for ESC, approx 10x20 mm

Or make some bigger redesign, like multiple parts and cable
could go out at the top. Split in half along side, or have a top-plate
Might want to consider millability at the same time then,
so it can be made on Modela, and/or in aluminium